# Speed up all modules by tackling two performance bottlenecks