--- a/multiqc/plots/linegraph.py
+++ b/multiqc/plots/linegraph.py
@@ -218,7 +218,7 @@ def smooth_line_data(data: Dict[str, Dict], numpoints: int) -> Dict[str, Dict[in
             continue
 
         binsize = (len(d) - 1) / (numpoints - 1)
-        first_element_indices = [round(binsize * i) for i in range(numpoints)]
+        first_element_indices = {round(binsize * i) for i in range(numpoints)}
         smoothed_d = {x: y for i, (x, y) in enumerate(d.items()) if i in first_element_indices}
         smoothed_data[s_name] = smoothed_d