- [x] This comment contains a description of changes (with reason)

I used the py-spy profiler. My previous PRs did not focus on the module code, as it looked like as if modules were slow or fast on their own accord. However I realized that some common code must be used because of the class inheritance. I did a more careful look at the profile and found a real stinker in the lineplot code that was triggered in multiple modules.

In a dict comprehension it was checked if something was a member of a list. List member checking time is proportional to the length of the list. I converted the list to a set. This one liner fix made a run on multiqc/test-data noticably faster. 

The other one I found was the `spectra.scale` call from the plotly library was expensive. I did some investigation of the input data and found that only 34 unique lists of colours were used, so I used a functools.lru_cache to simply cache the call. This also removed a further 10% of module runtime (on top of the previous savings). 

These are really minor changes codewise but have quite some benefit:

before:
```
|           multiqc |  - 9.59s: Searching files
|           multiqc |  - 8.41s: Running modules
|           multiqc |  - 5.04s: Compressing report data
```

After:
```
|           multiqc |  - 9.64s: Searching files
|           multiqc |  - 6.48s: Running modules
|           multiqc |  - 4.99s: Compressing report data
```

So a roughly 20 decrease in runtime for all module code.