@rhpvorderman 2024-04-27
Your welcome! We love to use MultiQC at our institute so I am simply trying to return in kind all the benefits you bring to us!