@ewels 2024-04-27
@rhpvorderman these PRs are fantastic, thank you!

When you multiply these speed ups by the number of runs that they will affect, it's a huge impact 👏🏻