@rhpvorderman 2024-04-29
The power to use profilers and change a few lines is insignficant next to the power of aggregating thousands of reports with MultiQC! I cannot thank you enough for making this great tool!

My main motivation is that MultiQC is so successful now and supports so many reports that scalability becomes a concern. Memory and runtimes tend to go up very steeply with increasing numbers of reports and this can lead to timeouts and out of memory errors on our SLURM cluster. Sure, more resources can be allocated, but the less the better. Also I have written a (very fast) FastQC/nanoplot replacement with Sequali, and I see MultiQC integration as a core feature. As such, if I improve the MultiQC user experience, I also improve that of Sequali users. It's a win-win basically.