@ewels 2024-04-29
Well, the more you can send our way the better! I have had a loose task of _"figure out how to use profilers properly and improve MultiQC performance"_ on my backlog for years, but it's one of those tasks that always gets buried because nothing's really actively broken or on fire or anything.

But totally agree that it's become more of an issue recently, also as sample numbers go up (looking at you, single cell data).